renamed auto-generated`.common` to `_common` across the board
