+++ api/.eslintignore
@@ -1,7 +1,7 @@
 node_modules/
 coverage
 dist
-src/.common
+src/_common
 oracle-cloud/build
 *.*
 !*.tsx
