+++ api/src/contribution/controller.ts
@@ -2,7 +2,7 @@ import { Controller, Get, QueryParams } from "routing-controllers";
 import {
   GetContributionsQueryDto,
   GetContributionsResponseDto,
-} from "../.common/types/api-responses";
+} from "../_common/types/api-responses";
 import { OpenAPI, ResponseSchema } from "routing-controllers-openapi";
 import { ContributionRepository } from "./repository";
 import { Service } from "typedi";
