+++ api/src/contributor/controller.ts
@@ -1,5 +1,5 @@
 import { Controller, Get, QueryParam } from "routing-controllers";
-import { GetContributorsResponseDto, GithubUserDto } from "../.common/types/api-responses";
+import { GetContributorsResponseDto, GithubUserDto } from "../_common/types/api-responses";
 import { OpenAPI, ResponseSchema } from "routing-controllers-openapi";
 import { GithubService } from "../github/service";
 import { Service } from "typedi";
