+++ api/src/github-user/controller.ts
@@ -1,6 +1,6 @@
 import { Controller, Get, Param } from "routing-controllers";
 import { OpenAPI, ResponseSchema } from "routing-controllers-openapi";
-import { GetUserResponseDto } from "../.common/types/api-responses";
+import { GetUserResponseDto } from "../_common/types/api-responses";
 import { GithubService } from "../github/service";
 import { Service } from "typedi";
