+++ data/.gitignore
@@ -1,5 +1,5 @@
 dist
-src/.common
+src/_common
 coverage
 
 # logs
