+++ data/package.json
@@ -4,7 +4,7 @@
   "description": "dzCode.io data code",
   "scripts": {
     "clear": "rimraf ./dist",
-    "common:sync": "rimraf ./src/.common && syncdir ../common/src ./src/.common -c",
+    "common:sync": "rimraf ./src/_common && syncdir ../common/src ./src/_common -c",
     "common:sync:watch": "yarn common:sync --watch",
     "lint:check": "yarn prettier:check && yarn eslint:check",
     "lint:fix": "yarn prettier:fix && yarn eslint:fix",
