+++ data/src/config/index.ts
@@ -1,3 +1,3 @@
-import { Environment } from "../.common/types";
-import { fsConfig } from "../.common/config";
+import { Environment } from "../_common/types";
+import { fsConfig } from "../_common/config";
 export const fullstackConfig = fsConfig(process.env as unknown as Environment);
