+++ data/src/index.ts
@@ -1,4 +1,4 @@
-import { getDataCollection, getDataEntry } from "./.common/utils/data";
+import { getDataCollection, getDataEntry } from "./_common/utils/data";
 import express from "express";
 import { fullstackConfig } from "./config";
 import { join } from "path";
