+++ web/.eslintignore
@@ -1,6 +1,6 @@
 node_modules/
 dist
-src/.common
+src/_common
 firebase
 coverage
 *.*
