+++ web/package.json
@@ -4,7 +4,7 @@
   "description": "frontend code for dzcode.io",
   "scripts": {
     "clear": "rimraf ./dist",
-    "common:sync": "rimraf ./src/.common && syncdir ../common/src ./src/.common -c",
+    "common:sync": "rimraf ./src/_common && syncdir ../common/src ./src/_common -c",
     "common:sync:watch": "yarn common:sync --watch",
     "dev": "yarn clear && cross-env NODE_ENV=development TS_NODE_PROJECT=\"webpack.tsconfig.json\" webpack serve --config ./webpack.config.ts",
     "lint:check": "yarn prettier:check && yarn eslint:check",
