+++ web/src/apps/main/pages/projects/catalog/index.tsx
@@ -1,7 +1,7 @@
 import { Card } from "src/apps/main/components/card";
 import { FC } from "react";
 import Grid from "@material-ui/core/Grid";
-import { Project } from "src/.common/types";
+import { Project } from "src/_common/types";
 import Typography from "@material-ui/core/Typography";
 import { makeStyles } from "@material-ui/core/styles";
