+++ web/src/apps/main/redux/actions/contribute-page/index.ts
@@ -1,6 +1,6 @@
 import Axios from "axios";
 import { ContributePageState } from "src/apps/main/redux/reducers/contribute-page";
-import { GetContributionsResponseDto } from "src/.common/types/api-responses";
+import { GetContributionsResponseDto } from "src/_common/types/api-responses";
 import { ThunkResult } from "src/apps/main/redux";
 import debounce from "@material-ui/core/utils/debounce";
 import { fullstackConfig } from "src/config";
