+++ web/src/apps/main/redux/reducers/articles-page/index.ts
@@ -1,5 +1,5 @@
 import { Action } from "src/apps/main/redux";
-import { Article } from "src/.common/types";
+import { Article } from "src/_common/types";
 import { SidebarTreeItem } from "src/apps/main/types";
 
 export interface ArticlesPageState {
