+++ web/src/config/index.ts
@@ -1,4 +1,4 @@
-import { fsConfig } from "../.common/config";
+import { fsConfig } from "../_common/config";
 import { getEnv } from "src/common/utils";
 
 export const fullstackConfig = fsConfig(getEnv(), {
